+++ src/imio/smartweb/core/viewlets/httpheaders.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from plone.app.layout.viewlets.httpheaders import HTTPCachingHeaders
+
+
+class MyHTTPCachingHeaders(HTTPCachingHeaders):
+    def update(self):
+        super(MyHTTPCachingHeaders, self).update()
+
+    def getHeaders(self):
+        result = super(MyHTTPCachingHeaders, self).getHeaders()
+        result.append(("prefix", "og: http://ogp.me/ns#"))
+        return result