+++ src/imio/smartweb/core/viewlets/ogptags.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from plone.app.layout.viewlets.httpheaders import HeaderViewlet
+
+
+class OgpTagsViewlet(HeaderViewlet):
+    def update(self):
+        super(OgpTagsViewlet, self).update()